pyproject.toml
[tool.poetry]
name = "latpm"
version = "0.1.0"
description = ""
authors = ["Mawwlle <mawwlle@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.27.1"
django-cors-headers = "^4.3.1"
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
ruff = "^0.2.2"
isort = "^5.13.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
